[core] Refactor duplicated code in Asset and File sources
