+++ platform/default/include/mbgl/storage/local_file_request.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <mbgl/util/string.hpp>
+
+namespace mbgl {
+
+template <typename>
+class ActorRef;
+class FileSourceRequest;
+
+void requestLocalFile(const std::string&, ActorRef<FileSourceRequest>);
+
+} // namespace mbgl
